requirements.txt
absl-py==0.13.0
anyascii==0.2.0
astunparse==1.6.3
autopep8==1.5.7
blis==0.7.4
CacheMan==2.1.0
cachetools==4.2.2
catalogue==2.0.5
certifi==2021.5.30
charset-normalizer==2.0.4
ci-info==0.2.0
clang==5.0
click==7.1.2
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
contractions==0.0.52
cycler==0.10.0
cyhunspell==2.0.2
cymem==2.0.5
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.1.0/en_core_web_lg-3.1.0-py3-none-any.whl
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
httplib2==0.19.1
idna==3.2
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
murmurhash==1.0.5
networkx==2.6.2
nibabel==3.2.1
nipype==1.6.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pathlib==1.0.1
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
protobuf==3.17.3
prov==2.0.0
psutil==5.8.0
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
pydot==1.4.2
PyMuPDF==1.18.17
pyparsing==2.4.7
pyspellchecker==0.6.2
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pyxnat==1.4
rdflib==6.0.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
simplejson==3.17.5
six==1.15.0
smart-open==5.1.0
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorhub==1.0.0a4
termcolor==1.1.0
textsearch==0.0.21
thinc==8.0.8
toml==0.10.2
tqdm==4.62.1
traits==6.2.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
